pyproject.toml
[project]
name = "auto-daily-report"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.1.7",
    "python-dotenv>=1.1.0",
]
